requirements.txt
openmc
numpy
matplotlib
pyyaml

results/sample/sample_enrichment_keff.csv